Use ncr_lists instead of list_sem_st in intefaces
Should result in no functionality change.

This makes the code marginally more effective (reducing the number of
"&lst->key" and "&lst->sessions" operations in the code, and moving them
toward dereferences where they can be combined with member accesses),
and more type-safe (prevents mixing the key and session list in most
places because they the difference is only in the low-level accessor
functions).

Most importantly, this allows replacing list_sem_st without having to
touch most of the functions again.