Use types from <linux/types.h> for public headers.
When included in the kernel tree, <inttypes.h> is not available, so
<linux/types.h> has to be used.  <linux/types.h> does not provide the
uintNN_t and size_t types when installed to /usr/include/linux (package
kernel-headers on Fedora).

Therefore, use the types from <linux/types.h> that are available in
userspace.

Also fix user-space users of the headers not to assume that they provide
the <stdint.h> types.