Fix compilation against linux 4.3
Linux Crypto API changed to using aead_request_set_ad()
instead of aead_request_set_assoc().

Based on Horia Geantă's patch from
https://mail.gna.org/public/cryptodev-linux-devel/2015-11/msg00002.html

However we need to version-guard the change for older versions of the
kernel to avoid the build from breaking for them.

Tested with 4.3 on a p1010rdb.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Reviewed-by: Horia Geantă <horia.geanta@freescale.com>
Signed-off-by: Phil Sutter <phil@nwl.cc>