fix clean-up on error path for crypto_create_session
This patch fixes clean-up on error path for failed allocations of
ses_new->pages or ses_new->sg. In these cases, allocations made in
cryptodev_hash_init have not been undone resulting in possible memory
leaks.

We take advantage of the initializations with zeros of the session
structure to trim the code to a single clean-up path.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>