fix destination for staged installs
The standard variable for staged installations is DESTDIR and it should
be set only by the user, outside of Makefile. This is consistent with
recommendations from both GNU Make and FreeBSD porters-handbook.

	make DESTDIR=/tmp/stage install

$prefix and $PREFIX (in FreeBSD world) are used to specify where the
files will be installed. It defaults to /usr/local but can be /usr or
/opt or something else.

Usually one wants to do a stage installation when cross-building or for
other purposes. The convention is to use DESTDIR.
Changing $prefix is useful for installers. They can override its value
when calling make:

	make prefix=/usr install

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>