avoid implicit conversion between signed and unsigned char
Use uint8_t type for all variables with this problem and avoid casting
from char in assignments. With uint8_t we also convey the information
that we're using small numbers rather than strings.

Although cryptodev.h uses the synonym type __u8, we use uint8_t
for consistency with other files in tests directory and also because it
is a standard POSIX type.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>