check session flags early to avoid incorrect failure modes
This verification of aead flag was incorrectly removed in
"refactoring: split big function to simplify maintainance"
20dcf071bc3076ee7db9d603cfbe6a06e86c7d5f
resulting in an incorrect dispatching of functions.

Add back this check and at the same time remove the second check from
the called function which now becomes redundant.
Add another guard check for aead modes and reject not supported combinations.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>