add CIOCCPHASH to copy hash state between sessions
It is often useful to copy the partial state of a hash function to
another session, so that it can be used as a base for two or more
computations.

The most prominent example is HMAC where all hashes start from
a base determined by the key.  By copying the state to another
session means that we only have to compute that base once rather
than for each message.

This creates a CIOCCPHASH ioctl, where the source and destination
session ids are fed to the paramenter struct.  Both sessions must have
been created before calling CIOCCPHASH.

The ioctl uses the kernel ahash import/export calls, available since
version 2.6.29.  The drivers were not required to implement these
calls until version 4.17.

OpenSSL requires this ioctl to avoid having to keep the contents of the
whole message before it can calculate the digest.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>