Compression: Add chunked mode
This commit extends compression so that the user can compress or decompress data in multiple separate chunks (compressed bitstreams) with a single ioctl system call. This avoids reducing the overhead if the user wants to compress the input in many small chunks, which previously required a separate ioctl for each chunk, involving substantial overhead.