Move recent Linux version #ifdefs from v6.4 to v6.5
The latest commits, meant to fix the build on Linux 6.4, are actually
fixing the build for API changes introduced in the merge window of the
yet-unreleased Linux 6.5, and actually break the build for Linux 6.4.

In particular, the upstream commits introducing the API changes are the
following, which are *not* included in the Linux v6.4 tag:
* https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=19c4e618a1bc3d0cad1f04c857be8076cb05bbb2
* https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=ca5e863233e8f6acd1792fd85d6bc2729a1b2c10

Change to #ifdef's to v6.5, where they will most likely be included.

Signed-off-by: Joan Bruguera Micó <joanbrugueram@gmail.com>