Exclude unused struct since Linux >= 6.5
Since Linux >= 6.5 we don't need `struct ctl_table verbosity_ctl_root`,
only the sysctl path, so move it to a constant and exclude the struct.

This is just a cosmetic change, intended to make it clear that neither
the struct nor the sentinel that follows are needed on newer kernels.

Signed-off-by: Joan Bruguera Micó <joanbrugueram@gmail.com>